+++ src/main/java/de/marhali/easyi18n/infra/properties/PropertiesWriter.java
@@ -6,6 +6,7 @@ import de.marhali.easyi18n.core.domain.model.I18nValue;
 import de.marhali.easyi18n.core.domain.model.TranslationConsumer;
 import de.marhali.easyi18n.core.domain.model.TranslationTarget;
 import de.marhali.easyi18n.core.domain.template.Templates;
+import de.marhali.easyi18n.core.ports.ProjectConfigPort;
 import de.marhali.easyi18n.infra.FileWriter;
 import org.jetbrains.annotations.NotNull;
 
@@ -22,8 +23,8 @@ public final class PropertiesWriter extends FileWriter {
 
     private final @NotNull LinkedProperties properties;
 
-    PropertiesWriter(@NotNull I18nPath path, @NotNull Templates templates) {
-        super(path, templates);
+    PropertiesWriter(@NotNull I18nPath path, @NotNull Templates templates, @NotNull ProjectConfigPort projectConfigPort) {
+        super(path, templates, projectConfigPort);
         this.properties = new LinkedProperties();
     }